+++ actionpack/lib/abstract_controller/helpers.rb
@@ -5,7 +5,6 @@ require "active_support/core_ext/name_error"
 
 module AbstractController
   module Helpers
-    include ActiveSupport::Deprecation::DeprecatedConstantAccessor
     extend ActiveSupport::Concern
 
     included do
@@ -23,23 +22,6 @@ module AbstractController
 
       self._helpers = define_helpers_module(self)
     end
-
-    class DeprecatedMissingHelperError < LoadError
-      def initialize(error, path)
-        @error = error
-        @path  = "helpers/#{path}.rb"
-        set_backtrace error.backtrace
-
-        if /^#{path}(\.rb)?$/.match?(error.path)
-          super("Missing helper file helpers/%s.rb" % path)
-        else
-          raise error
-        end
-      end
-    end
-    deprecate_constant "MissingHelperError", "AbstractController::Helpers::DeprecatedMissingHelperError",
-      message: "AbstractController::Helpers::MissingHelperError has been deprecated. If a Helper is not present, a NameError will be raised instead.",
-      deprecator: AbstractController.deprecator
 
     def _helpers
       self.class._helpers