+++ activejob/CHANGELOG.md
@@ -1,2 +1,17 @@
+*   Remove deprecated `:exponentially_longer` value for the `:wait` in `retry_on`.
+
+    *Rafael Mendonça França*
+
+*   Remove deprecated support to set numeric values to `scheduled_at` attribute.
+
+    *Rafael Mendonça França*
+
+*   Deprecate `Rails.application.config.active_job.use_big_decimal_serialize`.
+
+    *Rafael Mendonça França*
+
+*   Remove deprecated primitive serializer for `BigDecimal` arguments.
+
+    *Rafael Mendonça França*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/activejob/CHANGELOG.md) for previous changes.